+++ CHANGELOG.md
@@ -39,6 +39,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+* Added `create_class_assets` for invoke task to package the proto generated class file with `cpp`, `csharp` into assets.
+
 ### Changed
 
 * Fixed `compas_pb` support in Python 3.9.